packages/enterprise-constructs-base/pyproject.toml
[project]
name = "enterprise-constructs-base"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "aws-cdk-lib>=2.244.0,<3.0.0",
    "constructs>=10.5.0,<11.0.0",
]
